Move sweep runners + utils to anomalib
