+++ anomalib/utils/hpo/__init__.py
@@ -3,6 +3,6 @@
 # Copyright (C) 2022 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from .backends import CometSweep, WandbSweep
+from .runners import CometSweep, WandbSweep
 
 __all__ = ["CometSweep", "WandbSweep"]
